New type for getremoteactor

+++ packages/backend/utils/activitypub/processors/reject.ts
@@ -8,15 +8,17 @@ async function RejectActivity(body: activityPubObject, remoteUser: User, user: U
   const apObject: activityPubObject = body
   // someone rejected your follow request :(
   const userWichFollowWasRejected = await getRemoteActor(apObject.object.actor, user)
-  await Follows.destroy({
-    where: {
-      followedId: remoteUser.id,
-      followerId: userWichFollowWasRejected.id
-    }
-  })
-  redisCache.del('follows:full:' + userWichFollowWasRejected.id)
-  redisCache.del('follows:notYetAcceptedFollows:' + userWichFollowWasRejected.id)
-  // await signAndAccept({ body: body }, remoteUser, user)
+  if (userWichFollowWasRejected) {
+    await Follows.destroy({
+      where: {
+        followedId: remoteUser.id,
+        followerId: userWichFollowWasRejected.id
+      }
+    })
+    redisCache.del('follows:full:' + userWichFollowWasRejected.id)
+    redisCache.del('follows:notYetAcceptedFollows:' + userWichFollowWasRejected.id)
+    // await signAndAccept({ body: body }, remoteUser, user)
+  }
 }
 
 export { RejectActivity }
